problemas de arquitectura

+++ cases_backend/fastapi-backend/app/crud.py
@@ -1,6 +1,7 @@
 from sqlalchemy import asc, desc
 from sqlalchemy.orm import Session
 from . import models, schemas
+from sqlalchemy.orm import joinedload
 
 def create_aplicacion(db: Session, aplicacion: schemas.AplicacionCreate):
     db_aplicacion = models.Aplicacion(**aplicacion.dict())
@@ -12,8 +13,10 @@ def create_aplicacion(db: Session, aplicacion: schemas.AplicacionCreate):
 def get_aplicacion(db: Session, aplicacion_id: int):
     return db.query(models.Aplicacion).filter(models.Aplicacion.id == aplicacion_id).first()
 
-def get_aplicaciones(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc'):
+def get_aplicaciones(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc', filter_column: str = None, filter_value: str = None):
     query = db.query(models.Aplicacion)
+    if filter_column and filter_value:
+        query = query.filter(getattr(models.Aplicacion, filter_column).like(f'%{filter_value}%'))
     if sort_order == 'desc':
         query = query.order_by(desc(getattr(models.Aplicacion, sort_by)))
     else:
@@ -45,8 +48,11 @@ def create_pantalla(db: Session, pantalla: schemas.PantallaCreate):
 def get_pantalla(db: Session, pantalla_id: int):
     return db.query(models.Pantalla).filter(models.Pantalla.id == pantalla_id).first()
 
-def get_pantallas(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc'):
-    query = db.query(models.Pantalla)
+def get_pantallas(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc', filters: dict = {}):
+    query = db.query(models.Pantalla).join(models.Aplicacion)
+    if filters:
+        for key, value in filters.items():
+            query = query.filter(getattr(models.Pantalla, key).like(f'%{value}%'))
     if sort_order == 'desc':
         query = query.order_by(desc(getattr(models.Pantalla, sort_by)))
     else:
@@ -64,23 +70,36 @@ def update_pantalla(db: Session, pantalla_id: int, pantalla: schemas.PantallaUpd
     return db_pantalla
 
 def delete_pantalla(db: Session, pantalla_id: int):
-    db_pantalla = db.query(models.Pantalla).filter(models.Pantalla.id == pantalla_id).first()
-    db.delete(db_pantalla)
-    db.commit()
+    db_pantalla = db.query(models.Pantalla).options(
+        joinedload(models.Pantalla.aplicacion)
+    ).filter(models.Pantalla.id == pantalla_id).first()
+    if db_pantalla:
+        db.delete(db_pantalla)
+        db.commit()
     return db_pantalla
 
 def create_funcionalidad(db: Session, funcionalidad: schemas.FuncionalidadCreate):
-    db_funcionalidad = models.Funcionalidad(**funcionalidad.dict())
+    db_aplicacion = db.query(models.Aplicacion).filter(models.Aplicacion.id == funcionalidad.aplicacion_id).first()
+    db_funcionalidad = models.Funcionalidad(
+        nombre=funcionalidad.nombre,
+        aplicacion=db_aplicacion
+    )
     db.add(db_funcionalidad)
     db.commit()
     db.refresh(db_funcionalidad)
     return db_funcionalidad
 
 def get_funcionalidad(db: Session, funcionalidad_id: int):
-    return db.query(models.Funcionalidad).filter(models.Funcionalidad.id == funcionalidad_id).first()
+    return db.query(models.Funcionalidad).options(
+        joinedload(models.Funcionalidad.aplicacion)
+    ).filter(models.Funcionalidad.id == funcionalidad_id).first()
 
-def get_funcionalidades(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc'):
-    query = db.query(models.Funcionalidad)
+def get_funcionalidades(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc', filter_column: str = None, filter_value: str = None):
+    query = db.query(models.Funcionalidad).options(
+        joinedload(models.Funcionalidad.aplicacion)
+    )
+    if filter_column and filter_value:
+        query = query.filter(getattr(models.Funcionalidad, filter_column).like(f'%{filter_value}%'))
     if sort_order == 'desc':
         query = query.order_by(desc(getattr(models.Funcionalidad, sort_by)))
     else:
@@ -98,9 +117,12 @@ def update_funcionalidad(db: Session, funcionalidad_id: int, funcionalidad: sche
     return db_funcionalidad
 
 def delete_funcionalidad(db: Session, funcionalidad_id: int):
-    db_funcionalidad = db.query(models.Funcionalidad).filter(models.Funcionalidad.id == funcionalidad_id).first()
-    db.delete(db_funcionalidad)
-    db.commit()
+    db_funcionalidad = db.query(models.Funcionalidad).options(
+        joinedload(models.Funcionalidad.aplicacion)
+    ).filter(models.Funcionalidad.id == funcionalidad_id).first()
+    if db_funcionalidad:
+        db.delete(db_funcionalidad)
+        db.commit()
     return db_funcionalidad
 
 def create_so(db: Session, so: schemas.SOCreate):
@@ -111,10 +133,16 @@ def create_so(db: Session, so: schemas.SOCreate):
     return db_so
 
 def get_so(db: Session, so_id: int):
-    return db.query(models.SO).filter(models.SO.id == so_id).first()
+    return db.query(models.SO).options(
+        joinedload(models.SO.aplicacion)
+    ).filter(models.SO.id == so_id).first()
 
-def get_sos(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc'):
-    query = db.query(models.SO)
+def get_sos(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc', filter_column: str = None, filter_value: str = None):
+    query = db.query(models.SO).options(
+        joinedload(models.SO.aplicacion)
+    )
+    if filter_column and filter_value:
+        query = query.filter(getattr(models.SO, filter_column).like(f'%{filter_value}%'))
     if sort_order == 'desc':
         query = query.order_by(desc(getattr(models.SO, sort_by)))
     else:
@@ -132,9 +160,12 @@ def update_so(db: Session, so_id: int, so: schemas.SOUpdate):
     return db_so
 
 def delete_so(db: Session, so_id: int):
-    db_so = db.query(models.SO).filter(models.SO.id == so_id).first()
-    db.delete(db_so)
-    db.commit()
+    db_so = db.query(models.SO).options(
+        joinedload(models.SO.aplicacion)
+    ).filter(models.SO.id == so_id).first()
+    if db_so:
+        db.delete(db_so)
+        db.commit()
     return db_so
 
 def create_tipo_prueba(db: Session, tipo_prueba: schemas.TipoPruebaCreate):
@@ -145,10 +176,16 @@ def create_tipo_prueba(db: Session, tipo_prueba: schemas.TipoPruebaCreate):
     return db_tipo_prueba
 
 def get_tipo_prueba(db: Session, tipo_prueba_id: int):
-    return db.query(models.TipoPrueba).filter(models.TipoPrueba.id == tipo_prueba_id).first()
+    return db.query(models.TipoPrueba).options(
+        joinedload(models.TipoPrueba.aplicacion)
+    ).filter(models.TipoPrueba.id == tipo_prueba_id).first()
 
-def get_tipos_prueba(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc'):
-    query = db.query(models.TipoPrueba)
+def get_tipos_prueba(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc', filter_column: str = None, filter_value: str = None):
+    query = db.query(models.TipoPrueba).options(
+        joinedload(models.TipoPrueba.aplicacion)
+    )
+    if filter_column and filter_value:
+        query = query.filter(getattr(models.TipoPrueba, filter_column).like(f'%{filter_value}%'))
     if sort_order == 'desc':
         query = query.order_by(desc(getattr(models.TipoPrueba, sort_by)))
     else:
@@ -166,9 +203,12 @@ def update_tipo_prueba(db: Session, tipo_prueba_id: int, tipo_prueba: schemas.Ti
     return db_tipo_prueba
 
 def delete_tipo_prueba(db: Session, tipo_prueba_id: int):
-    db_tipo_prueba = db.query(models.TipoPrueba).filter(models.TipoPrueba.id == tipo_prueba_id).first()
-    db.delete(db_tipo_prueba)
-    db.commit()
+    db_tipo_prueba = db.query(models.TipoPrueba).options(
+        joinedload(models.TipoPrueba.aplicacion)
+    ).filter(models.TipoPrueba.id == tipo_prueba_id).first()
+    if db_tipo_prueba:
+        db.delete(db_tipo_prueba)
+        db.commit()
     return db_tipo_prueba
 
 def create_tipo_usuario(db: Session, tipo_usuario: schemas.TipoUsuarioCreate):
@@ -179,10 +219,17 @@ def create_tipo_usuario(db: Session, tipo_usuario: schemas.TipoUsuarioCreate):
     return db_tipo_usuario
 
 def get_tipo_usuario(db: Session, tipo_usuario_id: int):
-    return db.query(models.TipoUsuario).filter(models.TipoUsuario.id == tipo_usuario_id).first()
+    return db.query(models.TipoUsuario).options(
+        joinedload(models.TipoUsuario.aplicacion)
+    ).filter(models.TipoUsuario.id == tipo_usuario_id).first()
 
-def get_tipos_usuario(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc'):
-    query = db.query(models.TipoUsuario)
+
+def get_tipos_usuario(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc', filter_column: str = None, filter_value: str = None):
+    query = db.query(models.TipoUsuario).options(
+        joinedload(models.TipoUsuario.aplicacion)
+    )
+    if filter_column and filter_value:
+        query = query.filter(getattr(models.TipoUsuario, filter_column).like(f'%{filter_value}%'))
     if sort_order == 'desc':
         query = query.order_by(desc(getattr(models.TipoUsuario, sort_by)))
     else:
@@ -200,9 +247,12 @@ def update_tipo_usuario(db: Session, tipo_usuario_id: int, tipo_usuario: schemas
     return db_tipo_usuario
 
 def delete_tipo_usuario(db: Session, tipo_usuario_id: int):
-    db_tipo_usuario = db.query(models.TipoUsuario).filter(models.TipoUsuario.id == tipo_usuario_id).first()
-    db.delete(db_tipo_usuario)
-    db.commit()
+    db_tipo_usuario = db.query(models.TipoUsuario).options(
+        joinedload(models.TipoUsuario.aplicacion)
+    ).filter(models.TipoUsuario.id == tipo_usuario_id).first()
+    if db_tipo_usuario:
+        db.delete(db_tipo_usuario)
+        db.commit()
     return db_tipo_usuario
 
 def create_caso_prueba(db: Session, caso_prueba: schemas.CasoPruebaCreate):
@@ -213,10 +263,25 @@ def create_caso_prueba(db: Session, caso_prueba: schemas.CasoPruebaCreate):
     return db_caso_prueba
 
 def get_caso_prueba(db: Session, caso_prueba_id: int):
-    return db.query(models.CasoPrueba).filter(models.CasoPrueba.id == caso_prueba_id).first()
+    return db.query(models.CasoPrueba).options(
+        joinedload(models.CasoPrueba.funcionalidad),
+        joinedload(models.CasoPrueba.so),
+        joinedload(models.CasoPrueba.tipo_prueba),
+        joinedload(models.CasoPrueba.pantalla),
+        joinedload(models.CasoPrueba.aplicacion)
+    ).filter(models.CasoPrueba.id == caso_prueba_id).first()
 
-def get_casos_prueba(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc'):
-    query = db.query(models.CasoPrueba)
+
+def get_casos_prueba(db: Session, skip: int = 0, limit: int = 10, sort_by: str = 'id', sort_order: str = 'asc', filter_column: str = None, filter_value: str = None):
+    query = db.query(models.CasoPrueba).options(
+        joinedload(models.CasoPrueba.funcionalidad),
+        joinedload(models.CasoPrueba.so),
+        joinedload(models.CasoPrueba.tipo_prueba),
+        joinedload(models.CasoPrueba.pantalla),
+        joinedload(models.CasoPrueba.aplicacion)
+    )
+    if filter_column and filter_value:
+        query = query.filter(getattr(models.CasoPrueba, filter_column).like(f'%{filter_value}%'))
     if sort_order == 'desc':
         query = query.order_by(desc(getattr(models.CasoPrueba, sort_by)))
     else:
@@ -238,7 +303,14 @@ def update_caso_prueba(db: Session, caso_prueba_id: int, caso_prueba: schemas.Ca
     return db_caso_prueba
 
 def delete_caso_prueba(db: Session, caso_prueba_id: int):
-    db_caso_prueba = db.query(models.CasoPrueba).filter(models.CasoPrueba.id == caso_prueba_id).first()
-    db.delete(db_caso_prueba)
-    db.commit()
+    db_caso_prueba = db.query(models.CasoPrueba).options(
+        joinedload(models.CasoPrueba.funcionalidad),
+        joinedload(models.CasoPrueba.so),
+        joinedload(models.CasoPrueba.tipo_prueba),
+        joinedload(models.CasoPrueba.pantalla),
+        joinedload(models.CasoPrueba.aplicacion)
+    ).filter(models.CasoPrueba.id == caso_prueba_id).first()
+    if db_caso_prueba:
+        db.delete(db_caso_prueba)
+        db.commit()
     return db_caso_prueba
